Rename /config -> init in server/
